Refactor TextScriptParser into modular classes
